mcp-demo/requirements.txt
# MCP 基础库
# Model Context Protocol SDK
mcp==1.5.0
# HTTP 客户端库，用于 API 调用
httpx==0.27.0
requests>=2.31.0
# 异步库
asyncio==3.4.3
# OpenAI
openai==1.77.0
# LangChain MCP 适配器
langchain-mcp-adapters==0.0.9
langgraph==0.4.1
langchain-openai==0.3.16

# AWS OpenSearch MCP Server 依赖
fastmcp>=0.1.0
playwright>=1.40.0
python-dotenv>=1.0.0